+++ packages/br-validators/tests/types/validation-result.test.ts
@@ -4,6 +4,7 @@ import {
   brandCartaoCredito,
   brandCnpj,
   brandCep,
+  brandCnh,
   brandCpf,
   brandCodigoBarras,
   brandInscricaoEstadual,
@@ -25,6 +26,13 @@ describe('brandCpf', () => {
   it('casts string to Cpf brand', () => {
     const value = brandCpf('12345678909');
     expect(value).toBe('12345678909');
+  });
+});
+
+describe('brandCnh', () => {
+  it('casts string to Cnh brand', () => {
+    const value = brandCnh('62472927637');
+    expect(value).toBe('62472927637');
   });
 });